requirements.txt
discord.py>=2.3.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
aiohttp>=3.9.0

